Major changes
Deleted old openExperimentWIndow.mat file, and the group1_participant and group2_participant files previously used for counterbalancing. Added code in the IOGMotionMain file which counterbalances key presses with whether subjectNumber is even or odd. Made sure that the length of get.data.idDown is the same as the get.data.idUp for the csv file to be saved. Added formatResponses.mat file for key presses safety nets and saving keyboard info.